+++ docs/about/release-notes.rst
@@ -13,6 +13,11 @@ To upgrade Sdklib to the latest version, use pip:
 Sdklib 1.9.x series
 ===================
 
+Sdklib 1.9.3
+------------
+
+- Allow to ignore warnings.
+
 Sdklib 1.9.2
 ------------